+++ clang-tools-extra/clang-tidy/cert/CERTTidyModule.cpp
@@ -67,11 +67,6 @@ public:
     // MSC
     CheckFactories.registerCheck<LimitedRandomnessCheck>("cert-msc30-c");
   }
-  ClangTidyOptions getModuleOptions() override {
-    ClangTidyOptions Options;
-    Options.CheckOptions["cert-oop11-cpp.UseCERTSemantics"] = "1";
-    return Options;
-  }
 };
 
 } // namespace cert